refactor: PointGenerator 패키지 위치 util로 변경
