+++ backend/src/test/java/com/christmas/tree/domain/PointGeneratorTest.java
@@ -3,6 +3,7 @@ package com.christmas.tree.domain;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import com.christmas.util.PointGenerator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
